third_party/WebKit/LayoutTests/fast/dom/shadow/stylesheets-order-in-shadow-dom-expected.txt

CONSOLE WARNING: ::shadow pseudo-element is deprecated. See https://www.chromestatus.com/features/6750456638341120 for more details.
crbug.com/396585: ShadowDom CSS doesn't merge style

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(getNodeInComposedTree("host/world")).color is "rgb(255, 255, 0)"
PASS successfullyParsed is true

TEST COMPLETE
